[#175] refactor: 디바이스 패키지 구조 리팩토링
