+++ emulator/src/main/java/com/where_car/emulator/device/application/dto/CycleInfoDto.java
@@ -1,8 +1,10 @@
-package com.where_car.emulator.device.dto;
+package com.where_car.emulator.device.application.dto;
+
+import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.where_car.emulator.device.domain.common.CarCycleInfo;
-import java.util.List;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
