+++ emulator/src/main/java/com/where_car/emulator/device/infrastructure/JsonDatabase.java
@@ -1,19 +1,23 @@
-package com.where_car.emulator.device.repository;
+package com.where_car.emulator.device.infrastructure;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.DatabindException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.where_car.emulator.device.domain.common.CarIdentity;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Paths;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
+
 import javax.annotation.PostConstruct;
-import lombok.extern.slf4j.Slf4j;
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Repository;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.DatabindException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.where_car.emulator.device.domain.common.CarIdentity;
+
+import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Repository
